# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Interactive Fiction Platform API
  description: OpenAPI 3.0 Specification for Interactive Fiction Platform (DAG Novel System)
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local Development Server
paths:
...
  /api/me/novels:
    get:
      summary: Get novels authored by current user (Writer)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of writer novels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Novel'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /novels:
    get:
      summary: List all novels
      parameters:
        - in: query
          name: category_id
          schema:
            type: integer
        - in: query
          name: status
          schema:
            type: string
      responses:
        '200':
          description: List of novels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Novel'
    post:
      summary: Create new novel (Writer)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, description, category_id]
              properties:
                title:
                  type: string
                description:
                  type: string
                category_id:
                  type: integer
                cover_image:
                  type: string
      responses:
        '201':
          description: Novel created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Novel'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /novels/{id}:
    get:
      summary: Get novel detail by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Novel detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Novel'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Update novel detail (Writer)
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                category_id:
                  type: integer
                cover_image:
                  type: string
                status:
                  type: string
      responses:
        '200':
          description: Novel updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Novel'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete novel (Writer)
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Novel deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /novels/{id}/start:
    get:
      summary: Get root scene to start reading a novel
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Initial scene detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scene'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad Request Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource Not Found Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: BAD_REQUEST
            message:
              type: string
              example: Invalid input payload
...
    Novel:
      type: object
      properties:
        novel_id:
          type: integer
        title:
          type: string
        captions:
          type: string
        introduction:
          type: string
        cover_image:
          type: string
        status:
          type: string
        is_published:
          type: boolean
        is_completed:
          type: boolean
        is_banned:
          type: boolean
        ban_reason:
          type: string
        author_id:
          type: integer
        author_name:
          type: string
        pen_name:
          type: string
        views:
          type: integer
        chapter_count:
          type: integer
        scene_count:
          type: integer
        like_count:
          type: integer
        bookshelf_count:
          type: integer
...
    Scene:
      type: object
      properties:
        id:
          type: integer
        chapter_id:
          type: integer
        title:
          type: string
        content:
          type: string
        image_url:
          type: string
        is_ending:
          type: boolean
        ending_type:
          type: string
        ending_title:
          type: string
        ending_description:
          type: string
...